Allow negative continuation regex
When pretty-printing code introduces line breaks, it doesn't know its
own indentation level, so it's pretty common that the closing curly
brace of a data structure ends up on column 0.

This could be worked around with -c "^\w" but, for most log file
formats, it's safer to go with -C "^<re>" as usually there's a
timestamp there or other regex-friendly prefix.